src/requirements.txt
PyJWT
cryptography